pyproject.toml
[project]
name = "split_bot"
version = "1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "aiogram~=3.15.0",
    "aiogram-i18n~=1.4",
    "aiohttp~=3.10.8",
    "alembic~=1.14.0",
    "asyncpg~=0.29.0",
    "redis~=5.1.0",
    "sqlalchemy~=2.0.35",
    "msgspec~=0.18.6",
    "pydantic~=2.9.2",
    "pydantic-settings[yaml]~=2.6.0",
    "fluent_runtime~=0.4.0",
    "aiogram-contrib>=1.1.2",
    "stollen>=0.5.6",
    "pytonconnect-fixed==0.3.1",
    "pytoniq-core>=0.1.40",
    "qrcode>=8.0",
    "pillow>=11.0.0",
]

[tool.uv]
dev-dependencies = [
    "mypy~=1.12.0",
    "black~=24.10.0",
    "ruff~=0.7.1",
    "types-qrcode>=8.0.0.20241004",
]

[tool.black]
line-length = 99
exclude = "\\.?venv"

[tool.ruff]
target-version = "py38"
line-length = 99
lint.select = [
    "C",
    "DTZ",
    "E",
    "F",
    "I",
    "ICN",
    "ISC",
    "N",
    "PLC",
    "PLE",
    "Q",
    "T",
    "W",
    "YTT",
]
lint.ignore = ["N805"]
exclude = [
    ".venv",
    ".idea",
]

[tool.mypy]
plugins = [
    "sqlalchemy.ext.mypy.plugin",
    "pydantic.mypy"
]
exclude = [
    "venv",
    ".venv",
    ".idea",
    ".tests",
]
warn_unused_configs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_reexport = true
strict_equality = true
extra_checks = true

[[tool.mypy.overrides]]
module = [
    "redis.*",
    "pytonconnect.*",
    "pytoniq_core.*",
]
ignore_missing_imports = true
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["app.telegram.handlers.*"]
strict_optional = false
warn_return_any = false
disable_error_code = ["union-attr"]
